+++ l10n_br_fiscal/models/__init__.py
@@ -53,6 +53,5 @@ from . import res_country_state
 from . import operation_dashboard
 from . import subsequent_operation
 from . import subsequent_document
-from . import document_email
 from . import city_taxation_code
 from . import document_supplement
