+++ l10n_br_fiscal/models/document_type.py
@@ -35,12 +35,6 @@ class DocumentType(models.Model):
         required=True,
     )
 
-    document_email_ids = fields.One2many(
-        comodel_name="l10n_br_fiscal.document.email",
-        inverse_name="document_type_id",
-        string="Email Template Definition",
-    )
-
     document_serie_ids = fields.One2many(
         comodel_name="l10n_br_fiscal.document.serie",
         inverse_name="document_type_id",
