+++ l10n_br_fiscal/security/fiscal_security.xml
@@ -80,13 +80,4 @@
         >['|',('company_id','=',False),('company_id','in',company_ids)]</field>
     </record>
 
-    <record id="l10n_br_fiscal_document_email_rule" model="ir.rule">
-        <field name="name">Fiscal Document Email multi-company</field>
-        <field name="model_id" ref="model_l10n_br_fiscal_document_email" />
-        <field eval="True" name="global" />
-        <field
-            name="domain_force"
-        >['|',('company_id','=',False),('company_id','in',company_ids)]</field>
-    </record>
-
 </odoo>
