+++ l10n_br_fiscal/views/l10n_br_fiscal_menu.xml
@@ -569,15 +569,6 @@
         sequence="20"
     />
 
-    <!-- Fiscal Document Email -->
-    <menuitem
-        id="document_email_menu"
-        action="document_email_action"
-        groups="l10n_br_fiscal.group_manager"
-        parent="others_config_menu"
-        sequence="30"
-    />
-
     <!-- Comment -->
     <menuitem
         id="comment_menu"
