+++ l10n_br_fiscal_notification/__init__.py
@@ -1,0 +1,1 @@
+from . import models
