+++ l10n_br_fiscal_notification/models/__init__.py
@@ -1,0 +1,4 @@
+from . import res_company
+from . import document_type
+from . import document
+from . import document_email
