+++ l10n_br_fiscal_notification/readme/CONTRIBUTORS.md
@@ -1,0 +1,6 @@
+- [KMEE](https://www.kmee.com.br):
+  - Luis Felipe Mileo <mileo@kmee.com.br>
+- [Akretion](https://www.akretion.com/pt-BR):
+  - Renato Lima <renato.lima@akretion.com.br>
+- [Escodoo](https://www.escodoo.com.br):
+  - Marcel Savegnago <marcel.savegnago@escodoo.com.br>
