+++ l10n_br_fiscal_notification/security/fiscal_security.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo noupdate="1">
+
+    <record id="l10n_br_fiscal_document_email_rule" model="ir.rule">
+        <field name="name">Fiscal Document Email multi-company</field>
+        <field name="model_id" ref="model_l10n_br_fiscal_document_email" />
+        <field eval="True" name="global" />
+        <field
+            name="domain_force"
+        >['|',('company_id','=',False),('company_id','in',company_ids)]</field>
+    </record>
+
+</odoo>
