+++ l10n_br_fiscal_notification/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+"l10n_br_fiscal_document_email_user","Fiscal Document Email for User","model_l10n_br_fiscal_document_email","l10n_br_fiscal.group_user",1,0,0,0
+"l10n_br_fiscal_document_email_manager","Fiscal Document Email for Manager","model_l10n_br_fiscal_document_email","l10n_br_fiscal.group_manager",1,1,1,1
