+++ l10n_br_fiscal_notification/views/document_email_view.xml
