+++ l10n_br_fiscal_notification/views/document_type_view.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+    <record id="document_type_form" model="ir.ui.view">
+        <field name="model">l10n_br_fiscal.document.type</field>
+        <field name="inherit_id" ref="l10n_br_fiscal.document_type_form" />
+        <field name="arch" type="xml">
+            <notebook position="inside">
+                <page name="document_email" string="Email Templates">
+                    <field
+                        name="document_email_ids"
+                        context="{'default_document_type_id': id}"
+                    />
+                </page>
+            </notebook>
+        </field>
+    </record>
+
+</odoo>
