+++ l10n_br_fiscal_notification/views/document_view.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+  <record id="document_form" model="ir.ui.view">
+    <field name="model">l10n_br_fiscal.document</field>
+    <field name="inherit_id" ref="l10n_br_fiscal.document_form" />
+    <field name="arch" type="xml">
+            <button name="action_create_return" position="after">
+                    />
+                <button
+                    name="action_send_email"
+                    type="object"
+                    string="Send Email"
+                    groups="l10n_br_fiscal.group_user"
+                    attrs="{'invisible': [('state_edoc', 'not in', ('autorizada', 'cancelada', 'denegada'))]}"
+                />
+            </button>
+    </field>
+  </record>
+
+</odoo>
