+++ setup/l10n_br_fiscal_notification/odoo/addons/l10n_br_fiscal_notification
@@ -1,0 +1,1 @@
+../../../../l10n_br_fiscal_notification
